+++ app/WebsiteHeaders/HomeHeaders/home_research.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
 import { Container } from "react-bootstrap";
-import Research from "../WebsiteComponents/Research/research";
+import Research from "../../WebsiteComponents/Research/research";
 import { MdKeyboardArrowDown } from "react-icons/md";
 import { MdKeyboardArrowRight } from "react-icons/md";
