+++ app/WebsiteHeaders/HomeHeaders/home_skills.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
 import { Container } from "react-bootstrap";
-import Skills from "../WebsiteComponents/skills";
+import Skills from "../../WebsiteComponents/skills";
 import { MdKeyboardArrowDown } from "react-icons/md";
 import { MdKeyboardArrowRight } from "react-icons/md";
