--- app/WebsiteHeaders/SL_math100.tsx
@@ -1,6 +1,0 @@
-export default function SLMath100() {
-    return (
-        <>
-        </>
-    );
-}
