--- app/WebsiteHeaders/SL_math101.tsx
@@ -1,6 +1,0 @@
-export default function SLMath101() {
-    return (
-        <>
-        </>
-    );
-}
