--- app/WebsiteHeaders/contact_details
@@ -1,6 +1,0 @@
-export default function ContactDetails() {
-    return (
-        <>
-        </>
-    );
-}
