+++ app/page.tsx
@@ -1,11 +1,11 @@
-import Research from "./WebsiteHeaders/home_research";
-import Skills from "./WebsiteHeaders/home_skills";
-import Header from "./WebsiteHeaders/home_header"
-import Resume from "./WebsiteHeaders/home_resume"
-import Description from "./WebsiteHeaders/home_description";
-import Education from "./WebsiteHeaders/home_education";
-import WorkEx from "./WebsiteHeaders/home_workex";
-import Projects from "./WebsiteHeaders/home_projects";
+import Research from "./WebsiteHeaders/HomeHeaders/home_research";
+import Skills from "./WebsiteHeaders/HomeHeaders/home_skills";
+import Header from "./WebsiteHeaders/HomeHeaders/home_header"
+import Resume from "./WebsiteHeaders/HomeHeaders/home_resume"
+import Description from "./WebsiteHeaders/HomeHeaders/home_description";
+import Education from "./WebsiteHeaders/HomeHeaders/home_education";
+import WorkEx from "./WebsiteHeaders/HomeHeaders/home_workex";
+import Projects from "./WebsiteHeaders/HomeHeaders/home_projects";
 
 export default function Home() {
   return (
